Diary 6

+++ diary_app/Program.cs
@@ -1,14 +1,10 @@
-﻿using System;
-using System.Security.Cryptography.X509Certificates;
-using DiaryEntryNameSpace;
-using System.IO;
-using System.Reflection;
-using System.Diagnostics.Metrics;
+﻿using DiaryEntryNameSpace;
 
 
 namespace DiaryApp
 {
-    class Program {
+    class Program
+    {
 
         static void Main()
         {
@@ -25,14 +21,14 @@ namespace DiaryApp
             {
 
 
-                string answer = Console.ReadLine();
-                string aLow = answer.ToLower();
+                string? answer = Console.ReadLine();
+                string? aLow = answer?.ToLower();
 
                 switch (aLow)
                 {
                     case "a":
                         Console.WriteLine("");
-                        Add(); 
+                        Add();
                         break;
 
                     case "r":
@@ -65,12 +61,16 @@ namespace DiaryApp
             {
                 StreamWriter writer = File.AppendText("diary.txt");
                 Console.WriteLine("Please choose a title for your entry:");
-                string title = Console.ReadLine();
+                string? title = Console.ReadLine();
+                if (string.IsNullOrWhiteSpace(title) == true)
+                {
+                    title = "Default";
+                }
                 Console.WriteLine("\nEnter text your diary entry:");
-                string content = Console.ReadLine();
+                string? content = Console.ReadLine();
 
                 DiaryEntry entry = new DiaryEntry(title, content);
- 
+
                 writer.WriteLine(entry.Title + ";" + entry.Content + ";" + DateTime.Now);
                 writer.Close();
                 Console.WriteLine("\nEntry added!\n");
@@ -78,64 +78,69 @@ namespace DiaryApp
 
             void Read()
             {
-                Console.WriteLine("Please choose the title of the entry you would like to read:");
-                Console.WriteLine("(Note: If there are multiple entries with the same title, each one will be printed.)");
-
-                List();
-                var lines = File.ReadAllLines("diary.txt");
-                Console.WriteLine("");
-                string title = Console.ReadLine();
-                int id = Search(title);
-                if (id >= 0)
+                if (List() == true)
                 {
-                    Console.WriteLine("\n\n" + "\n" + lines[id].Split(";")[2] + "\n\n" + lines[id].Split(";")[0] + "\n" + lines[id].Split(";")[1]);
+                    Console.WriteLine("Please choose the title of the entry you would like to read:");
+                    Console.WriteLine("(Note: If there are multiple entries with the same title, each one will be printed.)");
+                    var lines = File.ReadAllLines("diary.txt");
+                    Console.WriteLine("");
+                    string? title = Console.ReadLine();
+                    int id = Search(title);
+                    if (id >= 0)
+                    {
+                        Console.WriteLine("\n\n" + lines[id].Split(";")[2] + "\n\n" + lines[id].Split(";")[0] + "\n" + lines[id].Split(";")[1] + "\n");
+                    }
                 }
             }
 
             void Modify()
             {
-                Console.WriteLine("Please choose the title of the entry you would like to modify:");
-                Console.WriteLine("(Note: If there are multiple entries with the same title, you can choose which one to modify.)");
-                List();
-                Console.WriteLine("");
-                string title = Console.ReadLine();
-                int id = Search(title);
-
-
-                Console.WriteLine("\nWarning: Modifying will overwrite existing title or content!");
-
-                if (Confirm() == true)
+                if (List() == true)
                 {
-                    ModifyEntry(id);
+                    Console.WriteLine("Please choose the title of the entry you would like to modify:");
+                    Console.WriteLine("(Note: If there are multiple entries with the same title, you can choose which one to modify.)");
+
+                    Console.WriteLine("");
+                    string? title = Console.ReadLine();
+                    int id = Search(title);
+
+
+                    Console.WriteLine("\nWarning: Modifying will overwrite existing title or content!");
+
+                    if (Confirm() == true)
+                    {
+                        ModifyEntry(id);
+                    }
                 }
             }
 
-            void ModifyEntry(int id) { 
+            void ModifyEntry(int id)
+            {
 
-                    var lines = File.ReadAllLines("diary.txt");
-                    var tempDiary = new StreamWriter("tempDiary.txt");
-                    while (true)
+                var lines = File.ReadAllLines("diary.txt");
+                var tempDiary = new StreamWriter("tempDiary.txt");
+                while (true)
+                {
+                    Console.WriteLine("What would you like to modify 't' for the title and 'c' for the content?");
+                    string? tOrC = Console.ReadLine();
+
+                    if (tOrC == "t")
                     {
-                        Console.WriteLine("What would you like to modify 't' for the title and 'c' for the content?");
-                        string tOrC = Console.ReadLine();
+                        Console.WriteLine("Please choose a new title: ");
+                        string? newTitle = Console.ReadLine();
+                        lines[id] = newTitle + ";" + lines[id].Split(";")[1] + ";" + lines[id].Split(";")[2];
+                        break;
+                    }
 
-                        if (tOrC == "t")
-                        {
-                            Console.WriteLine("Please choose a new title: ");
-                            string newTitle = Console.ReadLine();
-                            lines[id] = newTitle + ";" + lines[id].Split(";")[1] + ";" + lines[id].Split(";")[2];
-                            break;
-                        }
-
-                        if (tOrC == "c")
-                        {
-                            Console.WriteLine("Please write new content: ");
-                            string newContent = Console.ReadLine();
+                    if (tOrC == "c")
+                    {
+                        Console.WriteLine("Please write new content: ");
+                        string? newContent = Console.ReadLine();
                         lines[id] = lines[id].Split(";")[0] + ";" + newContent + ";" + lines[id].Split(";")[2];
                         break;
-                        }
-                        Console.WriteLine("Not a valid input, please try again.");
                     }
+                    Console.WriteLine("Not a valid input, please try again.");
+                }
 
                 for (int i = 0; i < lines.Length; i++)
                 {
@@ -145,61 +150,74 @@ namespace DiaryApp
                 File.Delete("diary.txt");
                 File.Move("tempDiary.txt", "diary.txt");
                 Console.WriteLine("\nModified!");
-                
+
             }
 
-                void Delete()
+            void Delete()
             {
-                Console.WriteLine("Please choose the title of the entry you would like to delete:");
-                Console.WriteLine("(Note: If there are multiple entries with the same title, you can choose which one to delete.)");
-                List();
-                Console.WriteLine("");
-                string title = Console.ReadLine();
 
-                int id = Search(title);
-
-                if (Confirm() == true)
+                if (List() == true)
                 {
-                    var lines = File.ReadAllLines("diary.txt");
-                    var tempDiary = new StreamWriter("tempDiary.txt");
+                    Console.WriteLine("Please choose the title of the entry you would like to delete:");
+                    Console.WriteLine("(Note: If there are multiple entries with the same title, you can choose which one to delete.)");
+                    Console.WriteLine("");
+                    string? title = Console.ReadLine();
+
+                    int id = Search(title);
+
+                    if (Confirm() == true)
+                    {
+                        var lines = File.ReadAllLines("diary.txt");
+                        var tempDiary = new StreamWriter("tempDiary.txt");
 
                         for (int i = 0; i < lines.Length; i++)
-                    {
-                        if (i != id)
                         {
-                            tempDiary.WriteLine(lines[i]);
+                            if (i != id)
+                            {
+                                tempDiary.WriteLine(lines[i]);
+                            }
+                            if (id == -1 && lines[i].Split(";")[0] != title)
+                            {
+                                tempDiary.WriteLine(lines[i]);
+                            }
                         }
-                        if (id == -1 && lines[i].Split(";")[0] != title)
-                        {
-                            tempDiary.WriteLine(lines[i]);
-                        }
+                        tempDiary.Close();
+                        File.Delete("diary.txt");
+                        File.Move("tempDiary.txt", "diary.txt");
+                        Console.WriteLine("\nDeleted!");
                     }
-                    tempDiary.Close();
-                    File.Delete("diary.txt");
-                    File.Move("tempDiary.txt", "diary.txt");
-                    Console.WriteLine("\nDeleted!");
                 }
             }
 
 
-            void List()
+            bool List()
             {
                 StreamReader reader = new StreamReader("diary.txt");
-                string line = "";
-
+                string? line = "";
+                int counter = 0;
                 while ((line = reader.ReadLine()) != null)
                 {
                     string[] splitted = line.Split(";");
                     Console.WriteLine(splitted[0]);
+                    counter++;
                 }
+                if (counter > 0)
+                {
+                    reader.Close();
+                    return true;
+                }
+                Console.WriteLine("No entries found!");
                 reader.Close();
+                return false;
             }
+
 
 
             int Search(string title)
             {
                 int id = -1;
                 int counter = 0;
+                var print = "";
 
                 var lines = File.ReadAllLines("diary.txt");
 
@@ -208,19 +226,26 @@ namespace DiaryApp
                     string[] splitted = lines[i].Split(";");
                     if (splitted[0].Equals(title))
                     {
-                        Console.WriteLine("\n\n" + "ID number:" + i + "\n" + splitted[2] + "\n\n" + splitted[0] + "\n");
+                        print += "\n\n" + "ID number:" + i + "\n" + splitted[2] + "\n\n" + splitted[0] + "\n";
                         counter++;
                         id = i;
                     }
                 }
                 if (counter > 1)
                 {
+                    Console.Write(print);
                     Console.WriteLine("\nWarning! Multiple entries with the same title.");
                     Console.WriteLine("Choose the corresponding ID number you would like to interact with.");
-                    if (int.TryParse(Console.ReadLine(), out int chosenId) && chosenId >= 0 && chosenId < lines.Length)
+                    while (true)
                     {
-                        Console.WriteLine("\nYou've chosen: " + chosenId + ".");
-                        id = chosenId;
+                        if (int.TryParse(Console.ReadLine(), out int chosenId) && chosenId >= 0 && chosenId < lines.Length && title == lines[chosenId].Split(";")[0])
+                        {
+                            Console.WriteLine("\nYou've chosen: " + chosenId + ".");
+                            id = chosenId;
+                            break;
+                        }
+                        Console.WriteLine("\nID and title doesn't match!");
+                        continue;
                     }
                 }
                 if (counter == 0)
@@ -236,7 +261,7 @@ namespace DiaryApp
                 while (true)
                 {
                     Console.WriteLine("'y' for YES and 'n' for NO");
-                    string yesOrNo = Console.ReadLine();
+                    string? yesOrNo = Console.ReadLine();
 
                     if (yesOrNo.ToLower() == "y")
                     {
